+++ api_demo/contract/Encryption.js
