+++ api_demo/contract/Encryption.php
