+++ api_demo/spot/Encryption.php
@@ -1,40 +1,58 @@
 <?php
 
-function get_signed_params($public_key, $secret_key, $params) {
-    $params = array_merge(['api_key' => $public_key], $params);
-    ksort($params);
-    //decode return value of http_build_query to make sure signing by plain parameter string
-    $signature = hash_hmac('sha256', urldecode(http_build_query($params)), $secret_key);
-    return http_build_query($params) . "&sign=$signature";
+$api_key='XXXXXXXX'; # Input your API Key
+$secret_key='XXXXXXXX'; # Input your Secret Key
+$url="https://api-testnet.bybit.com"; #Testnet Environment
+$curl = curl_init();
+
+function http_req($endpoint,$method,$params,$Info){
+    global $api_key, $secret_key, $url, $curl;
+    $timestamp = time() * 1000;
+    $params_for_signature= $timestamp . $api_key . "5000" . $params;
+    $signature = hash_hmac('sha256', $params_for_signature, $secret_key);
+    curl_setopt_array($curl, array(
+        CURLOPT_URL => $url . $endpoint,
+        CURLOPT_RETURNTRANSFER => true,
+        CURLOPT_ENCODING => '',
+        CURLOPT_MAXREDIRS => 10,
+        CURLOPT_TIMEOUT => 0,
+        CURLOPT_FOLLOWLOCATION => true,
+        CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
+        CURLOPT_CUSTOMREQUEST => $method,
+        CURLOPT_POSTFIELDS => $params,
+        CURLOPT_HTTPHEADER => array(
+          "X-BAPI-API-KEY: $api_key",
+          "X-BAPI-SIGN: $signature",
+          "X-BAPI-SIGN-TYPE: 2",
+          "X-BAPI-TIMESTAMP: $timestamp",
+          "X-BAPI-RECV-WINDOW: 5000",
+          "Content-Type: application/json"
+        ),
+      ));
+    echo $Info . "\n";
+    $response = curl_exec($curl);
+    echo $response . "\n";
 }
 
-$params = [
-	'symbol' => 'BITUSDT', 
-	'side' => 'Buy', 
-	'type' => 'Limit', 
-	'qty' => '10', 
-	'price' => '1', 
-	'time_in_force' => 'GoodTillCancel',
-	'reduce_only' => false,
-	'close_on_trigger' => false,
-	'timestamp' => time() * 1000
-];
+#Create Order
+$endpoint="/spot/v3/private/order";
+$method="POST";
+$orderLinkId=uniqid();
+$params='{"symbol":"BTCUSDT","orderType":"Limit","side":"Buy","orderLinkId":"' .  $orderLinkId . '","orderQty":"0.001","orderPrice":"10000","timeInForce":"GTC"}';
+http_req("$endpoint","$method","$params","Create Order");
 
-$url = 'https://api-testnet.bybit.com/spot/v1/order';
+#Get Order List
+$endpoint="/spot/v3/private/order";
+$method="GET";
+$params='orderLinkId=' . $orderLinkId;
+http_req($endpoint,$method,$params,"List Order");
 
-$public_key = 'XXXXXXXXXXXXXXXXXXXXXXX';
-$secret_key = 'XXXXXXXXXXXXXXXXXXXXXXX';
-$qs=get_signed_params($public_key, $secret_key, $params);
-$curl_url=$url."?".$qs;
-$curl=curl_init($curl_url);
-echo $curl_url . "\n";
-curl_setopt($curl, CURLOPT_URL, $curl_url);
-#curl_setopt($curl, CURLOPT_POSTFIELDS, $qs);
-curl_setopt($curl, CURLOPT_POST, true);
-curl_setopt($curl, CURLOPT_RETURNTRANSFER, 1);
-curl_setopt($curl, CURLOPT_FOLLOWLOCATION, 1);
-curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
-curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
-#curl_setopt($curl, CURLOPT_PROXY,"127.0.0.1:1087");
-$response=curl_exec($curl);
-echo $response . "\n";
+#Cancel Order
+$endpoint="/spot/v3/private/cancel-order";
+$method="POST";
+$params='{"orderLinkId":"' .  $orderLinkId .'"}';
+http_req($endpoint,$method,$params,"Cancel Order");
+
+curl_close($curl);
+
+
