+++ legacy_demo/api_demo/Encryption.cpp
