+++ legacy_demo/api_demo/Encryption.rb
