+++ legacy_demo/api_demo/asset/Encryption.js
