+++ legacy_demo/api_demo/asset/Encryption.php
