+++ legacy_demo/api_demo/encryption.sh
