+++ legacy_demo/api_demo/futures/Encryption.php
