+++ legacy_demo/api_demo/spot/Encryption.sh
