+++ legacy_demo/api_demo/usdc/Encryption.js
