+++ legacy_demo/api_demo/usdc/encryption.php
