+++ legacy_demo/wss_demo/js/ws_public_demo.js
